+++ resources/views/Supplier/list.blade.php
@@ -21,7 +21,7 @@
             <td>{{ $sp->supp_address }}</td>
             <td>{{ $sp->supp_contact }}</td>
             <td><a href="{{action('SupplierController@edit', $sp->id)}}" class="btn btn-default"><span class="glyphicon glyphicon-pencil"></span></a></td>
-            <td>{!! Form::open(['method' => 'DELETE','route' => ['supplier.destroy', $sp->id]]) !!}
+            <td>{!! Form::open(['method' => 'DELETE','route' => ['supplier.destroy', $sp->id], "onclick"=>"return confirm('Are you sure?')"]) !!}
                 {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
                 {!! Form::close() !!}</td>
             </tr>  @endforeach
